remove AdModel source files
