+++ src/containers_widget.h
@@ -21,6 +21,8 @@
 #define CONTAINERS_WIDGET_H
 
 #include "entry_widget.h"
+#include "ad_interface.h"
+#include "entry_model.h"
 
 class QItemSelection;
 class AdModel;
@@ -33,7 +35,7 @@ class ContainersWidget final : public EntryWidget {
 Q_OBJECT
 
 public:
-    ContainersWidget(AdModel *model, QWidget *parent);
+    ContainersWidget(AdModel *model_arg, QWidget *parent);
 
 signals:
     void selected_changed(const QString &dn);
@@ -41,9 +43,39 @@ signals:
 private slots:
     void on_selection_changed(const QItemSelection &selected, const QItemSelection &);
 
+    void on_ad_interface_login_complete(const QString &search_base, const QString &head_dn);
+    void on_attributes_changed(const QString &dn);
+    void on_delete_entry_complete(const QString &dn); 
+    void on_dn_changed(const QString &old_dn, const QString &new_dn);
+    void on_create_entry_complete(const QString &dn, NewEntryType type); 
+
 private:
+    AdModel *model = nullptr;
     EntryProxyModel *proxy = nullptr;
 
 };
+
+class AdModel final : public EntryModel {
+
+public:
+    enum Column {
+        Name,
+        DN,
+        COUNT,
+    };
+
+    enum Roles {
+        CanFetch = Qt::UserRole + 1,
+    };
+
+    explicit AdModel(QObject *parent);
+
+    bool canFetchMore(const QModelIndex &parent) const;
+    void fetchMore(const QModelIndex &parent);
+    bool hasChildren(const QModelIndex &parent) const override;
+
+};
+
+QIcon get_entry_icon(const QString &dn);
 
 #endif /* CONTAINERS_WIDGET_H */
